pyproject.toml
[project]
name = "gpt-ner-tagging"
version = "0.0.1"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "spacy (>=3.8.7,<4.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.33,<0.4.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "transformers (>=4.56.1,<5.0.0)",
    "retry (>=0.9.2,<0.10.0)",
    "datasets (>=4.1.1,<5.0.0)",
    "seqeval (>=1.2.2,<2.0.0)",
    "accelerate (>=1.10.1,<2.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
